Move every constructor inside the different versions companion objects
