--- core/src/main/scala/memeid/Constructors.scala
@@ -1,95 +1,0 @@
-package memeid
-
-import java.util.{UUID => JUUID}
-
-import scala.util.Try
-
-import memeid.JavaConverters._
-import memeid.bits._
-import memeid.digest._
-import memeid.node._
-import memeid.time._
-
-protected[memeid] object Mask {
-  val VERSION: Long = mask(4, 12)
-  val UB32: Long    = 0x00000000FFFFFFFFL
-
-  def version(msb: Long, version: Long): Long =
-    writeByte(VERSION, msb, version)
-}
-
-trait Constructors {
-
-  /**
-   * Creates a valid [[UUID]] from two [[Long]] values representing
-   * the most/least significant bits.
-   */
-  def from(msb: Long, lsb: Long): UUID = new JUUID(msb, lsb).asScala
-
-  /**
-   * Creates a [[memeid.UUID UUID]] from the [[java.util.UUID#toString string standard representation]]
-   * wrapped in a [[scala.util.Right Right]].
-   *
-   * Returns [[scala.util.Left Left]] with the error in case the string doesn't follow the
-   * string standard representation.
-   */
-  def from(s: String): Either[Throwable, UUID] = Try(JUUID.fromString(s).asScala).toEither
-
-  // Construct a v1 (time-based) UUID.
-  def v1(implicit N: Node, T: Time): UUID = {
-    val timestamp = T.monotonic
-    val low       = readByte(mask(32, 0), timestamp)
-    val mid       = readByte(mask(16, 32), timestamp)
-    val high      = readByte(mask(12, 48), timestamp)
-    val msb       = Mask.version(high, 1) | (low << 32) | (mid << 16)
-
-    val clkHigh = writeByte(mask(2, 6), readByte(mask(6, 8), N.id), 0x2)
-    val clkLow  = readByte(mask(8, 0), N.clockSequence.toLong)
-    val lsb     = writeByte(mask(8, 56), writeByte(mask(8, 48), N.id, clkLow), clkHigh)
-    new UUID.V1(new JUUID(msb, lsb))
-  }
-
-  // Construct a v4 (random) UUID.
-  def v4: UUID = new UUID.V4(JUUID.randomUUID)
-
-  // Construct a v4 (random) UUID from the given `msb` and `lsb`.
-  def v4(msb: Long, lsb: Long): UUID = {
-    val v4msb = writeByte(mask(4, 12), msb, 0x4)
-    val v4lsb = writeByte(mask(2, 62), lsb, 0x2)
-    new UUID.V4(new JUUID(v4msb, v4lsb))
-  }
-
-  // Construct a SQUUID (random, time-based) UUID.
-  def squuid(implicit T: Time): UUID = {
-    val uuid = v4
-
-    val timedMsb = (T.posix << 32) | (uuid.msb & Mask.UB32)
-
-    new UUID.V4(new JUUID(timedMsb, uuid.lsb))
-  }
-
-  def v3[A](namespace: UUID, local: A)(implicit D: Digestible[A]): UUID =
-    new UUID.V3(hashed(MD5, 3, namespace, local))
-
-  def v5[A](namespace: UUID, local: A)(implicit D: Digestible[A]): UUID =
-    new UUID.V5(hashed(SHA1, 5, namespace, local))
-
-  private def hashed[A: Digestible](
-      algo: Algorithm,
-      version: Long,
-      namespace: UUID,
-      local: A
-  ): JUUID = {
-    val digest = algo.digest
-    val ns     = Digestible[UUID].toByteArray(namespace)
-    digest.update(ns)
-    val name = Digestible[A].toByteArray(local)
-    digest.update(name)
-    val bytes  = digest.digest
-    val rawMsb = fromBytes(bytes.take(8))
-    val rawLsb = fromBytes(bytes.drop(8))
-    val msb    = Mask.version(rawMsb, version)
-    val lsb    = writeByte(mask(2, 52), rawLsb, 0x2)
-    new JUUID(msb, lsb)
-  }
-}
