+++ core/src/test/scala/memeid/SQUUIDSpec.scala
@@ -10,13 +10,13 @@ class SQUUIDSpec extends Specification with ScalaCheck with IOMatchers {
   "SQUUID constructor" should {
 
     "create version 4 UUIDs" in {
-      val uuids = (1 to 10).par.map(_ => UUID.squuid.version)
+      val uuids = (1 to 10).par.map(_ => UUID.V4.squuid.version)
 
       uuids.to[Set] must contain(exactly(4))
     }
 
     "not generate the same UUID twice" in {
-      val uuids = (1 to 10).par.map(_ => UUID.squuid)
+      val uuids = (1 to 10).par.map(_ => UUID.V4.squuid)
 
       uuids.to[Set].size must be equalTo 10
     }
